Move js tests for insert_online_link to new file
